+++ apps/weather_station/unabto_main.c
@@ -71,7 +71,7 @@ int main(int argc, char* argv[])
 #if (_POSIX_C_SOURCE >= 199309L)
         struct timespec sleepTime;
         sleepTime.tv_sec = 0;
-        sleepTime.tv_nsec = 10*1000000;
+        sleepTime.tv_nsec = (long)10*1000000;
         nanosleep(&sleepTime, NULL);
 #endif
     }
