+++ src/modules/random/openssl/unabto_random_openssl.c
@@ -6,7 +6,7 @@
 #include <openssl/rand.h>
 
 void nabto_random(uint8_t* buf, size_t len) {
-    if (!RAND_bytes((unsigned char*)buf, len)) {
+    if (!RAND_bytes((unsigned char*)buf, (int)len)) {
         NABTO_LOG_FATAL(("RAND_bytes failed."));
     }
 }
